requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
PGPy==0.6.0
web3==6.20.1  #optional for login attempts to Ethereum blockchain
psycopg[binary,pool]==3.2.1  # only needed if you use Postgres (SQLite works by default)
itsdangerous==2.2.0
PyJWT==2.9.0
python-gnupg